fix(cascade): move enums to everos.core.enums to satisfy import-linter
The layered-architecture contract forbids everos.memory from importing
everos.infra.  Move ChangeKind/ChangeType/ChangeStatus to
everos.core.enums — a shared layer accessible from both infra and memory.

This also fixes the integration test_lap_append_during_handler_no_loss
failure caused by the enum definitions living in a module that the
memory layer could not legally reach.
